+++ source/D2Hell/src/Archive.cpp
@@ -80,11 +80,11 @@ void __fastcall ARCHIVE_CloseFile(HD2ARCHIVE hArchive, HSFILE hFile)
  * 1.14c: Game.0x00514B87
  * D2XBeta: D2Server.dll.0x10009C76
  */
-size_t __fastcall ARCHIVE_GetFileSize(HD2ARCHIVE hArchive, HSFILE hFile, size_t* pdwFileSizeHigh)
+uint32_t __fastcall ARCHIVE_GetFileSize(HD2ARCHIVE hArchive, HSFILE hFile, uint32_t* pdwFileSizeHigh)
 {
 	D2_ASSERT(hFile != nullptr);
 
-	size_t dwFileSize = FOG_FGetFileSize(hFile, pdwFileSizeHigh);
+	uint32_t dwFileSize = FOG_FGetFileSize(hFile, pdwFileSizeHigh);
 	if (dwFileSize == 0)
 	{
 		char szArchivePath[MAX_PATH];
